Fixed a bug loading a lazy component
On the lazy component loading:
Even though the component was ES6, it was still necessary to import by calling default. Ref: https://webpack.js.org/migrate/4#import-and-commonjs

Disabled the web pack mode behaviour of setting the nodeEnv based on the set mode.
This is because we expect nodeEnv to have more than two possible values.

Upgraded preact compat. This was in hope to fix the lazy loading, but it probably didn’t have much effect. Still an upgrade should be better. Ref: https://github.com/developit/preact-compat/issues/484

Updated uglify JS for the same reason that updated preact compat.