Fixed issue where second and following scenarios would fail
Made sure the driver and browser stack local was created every time before each scenario, rather than reused.
This is because monster kills it (and creates it later) after every scenario.

Made sure local identifier was being set on browserstack local, so that it’s aligned with the driver which has it.

Removed a repeated require.

Removed the ability to have multiple nodes.