Feature/redux should not be a singleton cx 2050 (#881)
Move Redux and EventEmitter within the scope of our App component. In order to achieve this, I followed the example in the Redux documentation.
Partially fixed the issue with history navigation.
If two SDK instances are running at the same time the Redux store and events will work as expected. Unfortunately, there is still an issue with the history when clicking on the back button while 2 instances of the SDK are running at the same time, therefore two instances of the SDK should not be mounted in the same window.

With this change, the tearDown function will just remove the SDK node from the DOM.

Co-authored-by: Bart <ziembabartek@gmail.com>
Co-authored-by: Joao Trindade <trindade.joao@gmail.com>