Check glare result is sent before fetching it
The JS SDK was assuming that if a validation was present, the glare
validation was always present.

But when the image is a PDF, the JS SDK does not request a glare
validation.
This means that the API should not return a detect_glare validation
result.

But the API has been doing it (which is not coherent) and will continue
to do it for some time to maintain backwards compatibility.

But in spite of this, we can fix this error in the JS SDK, and to always
check if there is a glare result, before fetching it value.