Hide User Consent when consent already given
When User Consent is the first step, it wasn't hide when the consent was
already given.

We were not waiting for user consent to be injected, so the history
router was pointing to the wrong step indexes.

Now, we display the spinner until the user consent is injected, and then
we display history router with the proper list of steps. History Router
will also skip Steps flagged with `skip: true`.

Also, cleaning types to avoid the dependency between `StepsRouterProps`
and the Step Hook. And, adding a `loading` boolean instead of the
`status` variable.